# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FitnessApp.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/InBodyResult": {
      "get": {
        "tags": [
          "InBodyResult"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InBodyResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InBodyResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InBodyResult"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "InBodyResult"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InBodyResult"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InBodyResult"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InBodyResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InBodyResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InBodyResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InBodyResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/InBodyResult/{id}": {
      "get": {
        "tags": [
          "InBodyResult"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InBodyResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InBodyResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InBodyResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "InBodyResult"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InBodyResult"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InBodyResult"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InBodyResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "InBodyResult"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/InBodyResult/trends": {
      "get": {
        "tags": [
          "InBodyResult"
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "format": "double"
                    }
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "format": "double"
                    }
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "format": "double"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Progress"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProgressEntryModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProgressEntryModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddProgressEntryModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Progress/{metricType}": {
      "get": {
        "tags": [
          "Progress"
        ],
        "parameters": [
          {
            "name": "metricType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Progress/{id}": {
      "put": {
        "tags": [
          "Progress"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressEntryModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressEntryModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressEntryModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Progress"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddProgressEntryModel": {
        "type": "object",
        "properties": {
          "metricType": {
            "type": "string",
            "nullable": true
          },
          "metricValue": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApplicationUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "profileImage": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "inBodyResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InBodyResult"
            },
            "nullable": true
          },
          "workoutPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutPlan"
            },
            "nullable": true
          },
          "completedWorkouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedWorkout"
            },
            "nullable": true
          },
          "userGoal": {
            "$ref": "#/components/schemas/UserGoal"
          },
          "progressEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserProgress"
            },
            "nullable": true
          },
          "resetPasswordToken": {
            "type": "string",
            "nullable": true
          },
          "resetPasswordTokenExpiry": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CompletedWorkout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "workoutPlanId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "caloriesBurned": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "intensity": {
            "type": "string",
            "nullable": true
          },
          "completionPercentage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "workoutPlan": {
            "$ref": "#/components/schemas/WorkoutPlan"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExercisePerformance"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Exercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "muscleGroup": {
            "type": "string",
            "nullable": true
          },
          "equipment": {
            "type": "string",
            "nullable": true
          },
          "difficultyLevel": {
            "type": "string",
            "nullable": true
          },
          "instructionSteps": {
            "type": "string",
            "nullable": true
          },
          "videoUrl": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "aiModelReference": {
            "type": "string",
            "nullable": true
          },
          "workoutExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExercisePerformance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "completedWorkoutId": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "weight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "reps": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sets": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "formQuality": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "formFeedback": {
            "type": "string",
            "nullable": true
          },
          "completedWorkout": {
            "$ref": "#/components/schemas/CompletedWorkout"
          },
          "exercise": {
            "$ref": "#/components/schemas/Exercise"
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false
      },
...
      "InBodyResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "bodyFatPercentage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "muscleMass": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bmi": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "visceralFat": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bmr": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tdee": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "activityLevel": {
            "type": "string",
            "nullable": true
          },
          "tdeeLastCalculated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProgressEntryModel": {
        "type": "object",
        "properties": {
          "metricValue": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserGoal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "goalType": {
            "type": "string",
            "nullable": true
          },
          "targetValue": {
            "type": "number",
            "format": "double"
          },
          "startValue": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "targetDate": {
            "type": "string",
            "format": "date-time"
          },
          "achieved": {
            "type": "boolean"
          },
          "dateAchieved": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false
      },
...
      "UserProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "metricType": {
            "type": "string",
            "nullable": true
          },
          "metricValue": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false
      },
      "WorkoutDay": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "workoutPlanId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "workoutDayId": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "repsPerSet": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "orderInWorkout": {
            "type": "integer",
            "format": "int32"
          },
          "restBetweenSets": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "workoutDay": {
            "$ref": "#/components/schemas/WorkoutDay"
          },
          "exercise": {
            "$ref": "#/components/schemas/Exercise"
          }
        },
        "additionalProperties": false
      },
      "WorkoutPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "isTemplate": {
            "type": "boolean"
          },
          "difficultyLevel": {
            "type": "string",
            "nullable": true
          },
          "focusArea": {
            "type": "string",
            "nullable": true
          },
          "durationWeeks": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "workoutDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutDay"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}